namespace実装
