+++ src/drag_and_drop_app.ts
@@ -1,381 +1,32 @@
-// Drag & Drop
-interface Draggable {
-    dragStartHandler(event: DragEvent): void;
-    dragEndHandler(event: DragEvent): void;
-}
+// import { Draggable, DragTarget } from "./models/drag-drop.js"
 
-interface DragTarget {
-    dragOverHandler(event: DragEvent): void;
-    dropHandler(event: DragEvent): void;
-    dragLeaveHandler(event: DragEvent): void;
-}
+/// <reference path="./models/drag-drop.ts" />
+/// <reference path="./models/project.ts" />
+/// <reference path="./state/project-state.ts" />
+/// <reference path="./util/validation.ts" />
+/// <reference path="./decorators/autobind.ts" />
+/// <reference path="./components/base-components.ts" />
+/// <reference path="./components/project-input.ts" />
+/// <reference path="./components/project-item.ts" />
+/// <reference path="./components/project-list.ts" />
 
-// Project Type
-enum ProjectStatus {
-    Active, Finished
-}
+namespace App {
 
-class Project {
-    constructor(
-        public id: string,
-        public title: string,
-        public description: string,
-        public manday: number,
-        public status: ProjectStatus,
-    ) { }
+    // Drag & Drop
+    // interface Draggable {
+    //     dragStartHandler(event: DragEvent): void;
+    //     dragEndHandler(event: DragEvent): void;
+    // }
 
-}
-
-// Project State Management
-// import { v4 as uuidv4 } from "uuid"; //なぜかuuidが使えなかった。。原因突き止められず
-type Listener<T> = (items: T[]) => void;
-
-abstract class State<T> {
-    protected listeners: Listener<T>[] = [];
-
-    addListener(listenerFn: Listener<T>) {
-        this.listeners.push(listenerFn);
-    }
-}
+    // interface DragTarget {
+    //     dragOverHandler(event: DragEvent): void;
+    //     dropHandler(event: DragEvent): void;
+    //     dragLeaveHandler(event: DragEvent): void;
+    // }
 
 
-class ProjectState extends State<Project> {
-    private projects: Project[] = [];
-    private static instance: ProjectState;
-
-    private constructor() { //シングルトーンにする場合にプライベートのコンストラクターを生成。外からインスタンス生成ができないように！！
-        super()
-    }
-
-    static getInstance() {
-        if (this.instance) {
-            return this.instance;
-        }
-        this.instance = new ProjectState();
-        return this.instance;
-    }
-
-
-
-    addProject(title: string, description: string, manday: number) {
-        const newProject = new Project(
-            Math.random().toString(),
-            title,
-            description,
-            manday,
-            ProjectStatus.Active,
-        )
-        this.projects.push(newProject);
-        this.updateListeners();
-    }
-
-    moveProject(projectId: string, newStatus: ProjectStatus) {
-        const project = this.projects.find(prj => prj.id === projectId);
-        if (project && project.status !== newStatus) {
-            project.status = newStatus;
-            this.updateListeners();
-        }
-    }
-
-    private updateListeners() {
-        for (const listenerFn of this.listeners) {
-            listenerFn(this.projects.slice()); //オリジナルを渡すと参照先を渡すため参照先で変更可能。そのため、あちこちで変更したくないときはコピーで渡す
-        }
-    }
+    const prjInput = new ProjectInput();
+    const activePjtList = new ProjectList("active");
+    const finishedPjtList = new ProjectList("finished");
 
 }
-const projectState = ProjectState.getInstance(); //シングルトーン　アプリ内で１つのインスタンスしか存在しないことを保証
-
-// validation
-interface Validatable {
-    value: string | number;
-    required?: boolean;
-    minLength?: number;
-    maxLength?: number;
-    min?: number;
-    max?: number;
-}
-function validate(validatableInput: Validatable) {
-    let isValid = true;
-    if (validatableInput.required) {
-        isValid = isValid && validatableInput.value.toString().trim().length !== 0;
-    }
-    if (validatableInput.minLength != null &&
-        typeof validatableInput.value === "string"
-    ) {
-        isValid =
-            isValid && validatableInput.value.length >= validatableInput.minLength;
-    }
-    if (validatableInput.maxLength != null &&
-        typeof validatableInput.value === "string"
-    ) {
-        isValid =
-            isValid && validatableInput.value.length <= validatableInput.maxLength;
-    }
-    if (validatableInput.min != null &&
-        typeof validatableInput.value === "number"
-    ) {
-        isValid =
-            isValid && validatableInput.value >= validatableInput.min;
-    }
-    if (validatableInput.max != null &&
-        typeof validatableInput.value === "number"
-    ) {
-        isValid =
-            isValid && validatableInput.value <= validatableInput.max;
-    }
-    return isValid;
-}
-
-// autobind decorator
-function autobind(
-    _: any,
-    _2: string,
-    descriptor: PropertyDescriptor
-) {
-    const originalMethod = descriptor.value;
-    const adjDescriptor: PropertyDescriptor = {
-        configurable: true, //propertyを変更できるようにするため
-        get() {
-            const boundFn = originalMethod.bind(this);
-            return boundFn;
-        }
-    }
-    return adjDescriptor;
-}
-
-// Compornent Class
-abstract class Compornent<T extends HTMLElement, U extends HTMLElement> {
-    templateElement: HTMLTemplateElement;
-    hostElement: T;
-    element: U;
-
-    constructor(
-        templateId: string,
-        hostElementId: string,
-        insertAtStart: boolean,
-        newElementId?: string
-    ) {
-        this.templateElement = document.getElementById(templateId)! as HTMLTemplateElement;
-        this.hostElement = document.getElementById(hostElementId)! as T;
-
-        const importedNode = document.importNode(this.templateElement.content, true);
-        this.element = importedNode.firstElementChild as U;
-        if (newElementId) {
-            this.element.id = newElementId;
-        }
-
-        this.attach(insertAtStart);
-    }
-
-    abstract configure(): void;
-    abstract renderContent(): void;
-
-    private attach(insertAtBeginning: boolean) {
-        this.hostElement.insertAdjacentElement(
-            insertAtBeginning ? "afterbegin" : "beforeend",
-            this.element
-        );
-    }
-}
-
-// ProjectItem Class
-class ProjectItem extends Compornent<HTMLUListElement, HTMLLIElement>
-    implements Draggable { // implements : Draggableの「型を持つことを約束する」という意味
-    private project: Project;
-
-    get manday() {
-        if (this.project.manday < 20) {
-            return `${this.project.manday.toString()}人日`
-        } else {
-            return `${(this.project.manday / 20).toString()}人月`
-        }
-    }
-
-    constructor(hostId: string, project: Project) {
-        super("single-project", hostId, false, project.id);
-        this.project = project;
-
-        this.configure();
-        this.renderContent();
-    }
-
-    @autobind
-    dragStartHandler(event: DragEvent) {
-        event.dataTransfer!.setData("text/plain", this.project.id);
-        event.dataTransfer!.effectAllowed = "move";
-    }
-
-    @autobind
-    dragEndHandler(_: DragEvent) {
-        console.log("Drag終了");
-    }
-
-    configure() {
-        this.element.addEventListener("dragstart", this.dragStartHandler);
-        this.element.addEventListener("dragend", this.dragEndHandler);
-    }
-    renderContent() {
-        this.element.querySelector("h2")!.textContent = this.project.title;
-        this.element.querySelector("h3")!.textContent = this.manday; //getter, setterはプロパティーのようにふるまうので、関数演算子の()は不要
-        this.element.querySelector("p")!.textContent = this.project.description;
-    }
-}
-
-
-// ProjectList Class
-class ProjectList extends Compornent<HTMLDivElement, HTMLElement>
-    implements DragTarget {
-    assignedProjects: Project[];
-
-    constructor(private type: "active" | "finished") {
-        super("project-list", "app", false, `${type}-projects`)
-        this.assignedProjects = [];
-
-        this.configure();
-        this.renderContent();
-    }
-
-    @autobind
-    dragOverHandler(event: DragEvent) {
-        if (event.dataTransfer && event.dataTransfer.types[0] === "text/plain") {
-            event.preventDefault(); //これをしないと、ドロップイベントが発生しない。addEventListenerで監視できない
-            const listEl = this.element.querySelector("ul")!;
-            listEl.classList.add("droppable");
-
-        }
-    }
-
-    @autobind
-    dropHandler(event: DragEvent) {
-        const prjId = event.dataTransfer!.getData("text/plain");
-        projectState.moveProject(
-            prjId, this.type === "active" ?
-            ProjectStatus.Active
-            :
-            ProjectStatus.Finished)
-    }
-
-    @autobind
-    dragLeaveHandler(_: DragEvent) {
-        const listEl = this.element.querySelector("ul")!;
-        listEl.classList.remove("droppable");
-    }
-
-    configure() {
-        this.element.addEventListener("dragover", this.dragOverHandler);
-        this.element.addEventListener("drop", this.dropHandler);
-        this.element.addEventListener("dragleave", this.dragLeaveHandler);
-
-        projectState.addListener((projects: Project[]) => {
-            const relevantProjects = projects.filter((prj) => {
-                if (this.type === "active") {
-                    return prj.status === ProjectStatus.Active;
-                }
-                return prj.status === ProjectStatus.Finished;
-            })
-            this.assignedProjects = relevantProjects;
-            this.renderProjects();
-        })
-    }
-
-    renderContent() {
-        const listId = `${this.type}-projects-list`;
-        this.element.querySelector("ul")!.id = listId;
-        this.element.querySelector("h2")!.textContent =
-            this.type === "active" ? "実行中プロジェクト" : "完了プロジェクト";
-    }
-
-    private renderProjects() {
-        const listEl = document.getElementById(`${this.type}-projects-list`)! as HTMLUListElement;
-        listEl.innerHTML = "";
-        for (const pjtItem of this.assignedProjects) {
-            new ProjectItem(listEl.id, pjtItem);
-        }
-    }
-}
-
-
-// ProjectInput Class
-class ProjectInput extends Compornent<HTMLDivElement, HTMLFormElement>{
-    titleInputElement: HTMLInputElement;
-    descriptionInputElement: HTMLInputElement;
-    mandayInputElement: HTMLInputElement;
-
-    constructor() {
-        super("project-input", "app", true, "user-input");
-
-        this.titleInputElement = this.element.querySelector("#title") as HTMLInputElement;
-        this.descriptionInputElement = this.element.querySelector("#description") as HTMLInputElement;
-        this.mandayInputElement = this.element.querySelector("#manday") as HTMLInputElement;
-
-        this.configure();
-    }
-
-    configure() {
-        // this.element.addEventListener("submit", this.submitHandler.bind(this)); bindメソッドで、submitHndler側のthisもこのクラスのオブジェクトを指すことを明示的に示す。他のやり方はデコレーターを使う
-        this.element.addEventListener("submit", this.submitHandler);
-    }
-
-    renderContent() { }
-
-    private gatherUserInput(): [string, string, number] | void { //union型でvoid型を入れることでalartのようにタプルを返さない条件分岐もOKにする。
-        const enteredTitle = this.titleInputElement.value;
-        const enteredDescription = this.descriptionInputElement.value;
-        const enteredManday = this.mandayInputElement.value;
-
-        const titleValidatable: Validatable = {
-            value: enteredTitle,
-            required: true,
-        }
-        const descriptionValidatable: Validatable = {
-            value: enteredDescription,
-            required: true,
-            minLength: 5,
-        }
-        const mandayValidatable: Validatable = {
-            value: parseFloat(enteredManday),
-            required: true,
-            min: 1,
-            max: 1000,
-        }
-
-        if ( // validationを設定する方法
-            // enteredTitle.trim().length === 0 ||
-            // enteredDescription.trim().length === 0 ||
-            // enteredManday.trim().length === 0
-            !validate(titleValidatable) ||
-            !validate(descriptionValidatable) ||
-            !validate(mandayValidatable)
-        ) {
-            alert("入力値が正しくありません。再度お試しください");
-        } else {
-            return [enteredTitle, enteredDescription, parseFloat(enteredManday)];
-        }
-    }
-
-    private clearInputs() {
-        this.titleInputElement.value = "";
-        this.descriptionInputElement.value = "";
-        this.mandayInputElement.value = "";
-    }
-
-    @autobind
-    private submitHandler(event: Event) {
-        event.preventDefault();
-        const userInput = this.gatherUserInput();
-        if (Array.isArray(userInput)) {
-            const [title, desc, manday] = userInput;
-            projectState.addProject(title, desc, manday);
-            // console.log(title, desc, manday);
-            this.clearInputs();
-        }
-    }
-
-
-}
-
-
-const prjInput = new ProjectInput();
-const activePjtList = new ProjectList("active");
-const finishedPjtList = new ProjectList("finished");
