+++ src/study/generic.ts
